+++ packages/wallet/src/initialization/instances/gas-fee-controller/gas-fee-controller.test.ts
@@ -217,9 +217,6 @@ function getInstanceOptions(): WalletOptions['instanceOptions'] {
       storage: new InMemoryStorageAdapter(),
     },
     remoteFeatureFlagController: REMOTE_FEATURE_FLAG_OPTIONS,
-    subscriptionService: {
-      fetchFunction: globalThis.fetch,
-    },
   };
 }